Reboot_Dana/requirements.txt
# AlphaEarth Crop Recommender Requirements

# Core ML and Data Science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
pickle-mixin>=1.0.2

# Google Earth Engine
earthengine-api>=0.1.300

# Web Framework
Flask>=2.0.0
Werkzeug>=2.0.0

# Async Support
asyncio-mqtt>=0.11.0

# Utilities
requests>=2.25.0
python-dotenv>=0.19.0

# Azure OpenAI for Agricultural Advice
openai>=1.0.0

# Development and Testing
pytest>=6.0.0
pytest-asyncio>=0.18.0

# Optional: Enhanced Features
# redis>=4.0.0  # For caching
# psycopg2>=2.9.0  # For PostgreSQL
# gunicorn>=20.0.0  # For production deployment